fix(striker-ui): disable network mapping on successful submission
