+++ striker-ui/components/StrikerInitForm.tsx
@@ -27,6 +27,7 @@ import NetworkInitForm, {
   NetworkInitFormValues,
 } from './NetworkInitForm';
 import { Panel, PanelHeader } from './Panels';
+import setMapNetwork from '../lib/setMapNetwork';
 import Spinner from './Spinner';
 import { BodyText, HeaderText, InlineMonoText, MonoText } from './Text';
 import useProtectedState from '../hooks/useProtectedState';
@@ -303,6 +304,9 @@ const StrikerInitForm: FC = () => {
           api
             .put('/init', requestBody)
             .then(() => {
+              // Stop network mapping only on successful form submission.
+              setMapNetwork(0);
+
               setIsSubmittingForm(false);
               setSubmitMessage({
                 children: (
